Make the app usable without a database connection and configure tests/CI

+++ components/AdvertisementSlideshow.tsx
@@ -42,14 +42,18 @@ const Slideshow = () => {
             color="secondary"
             onClick={async (e) => {
               e.preventDefault();
-              const response = await fetch("/api/advertisements", {
-                method: "DELETE",
-                headers: {
-                  "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ id: state.advertisements[index].id }),
-              }).then((response) => response.json());
-              if (response) {
+              let response;
+              if (process.env.DATABASE_URL) {
+                response = await fetch("/api/advertisements", {
+                  method: "DELETE",
+                  headers: {
+                    "Content-Type": "application/json",
+                  },
+                  body: JSON.stringify({ id: state.advertisements[index].id }),
+                }).then((response) => response.json());
+              }
+
+              if (response || !process.env.DATABASE_URL) {
                 dispatch({
                   type: "REMOVE_ADVERTISEMENT",
                   payload: state.advertisements[index].id,
